+++ mml/examples/pgp-commands.rs
@@ -2,19 +2,19 @@
 #[tokio::main]
 async fn main() {
     use mml::{
-        pgp::{CmdsPgp, Pgp},
+        pgp::{Pgp, PgpCommands},
         MmlCompilerBuilder,
     };
     use process::Command;
 
     let mml = include_str!("./pgp.eml");
     let mml_compiler = MmlCompilerBuilder::new()
-        .with_pgp(Pgp::Cmds(CmdsPgp {
+        .with_pgp(Pgp::Commands(PgpCommands {
             encrypt_cmd: Some(Command::from(
                 "gpg --homedir ./tests/gpg-home -eqa <recipients>",
             )),
-            encrypt_recipient_fmt: Some(CmdsPgp::default_encrypt_recipient_fmt()),
-            encrypt_recipients_sep: Some(CmdsPgp::default_encrypt_recipients_sep()),
+            encrypt_recipient_fmt: Some(PgpCommands::default_encrypt_recipient_fmt()),
+            encrypt_recipients_sep: Some(PgpCommands::default_encrypt_recipients_sep()),
             decrypt_cmd: Some(Command::from("gpg --homedir ./tests/gpg-home -dq")),
             sign_cmd: Some(Command::from("gpg --homedir ./tests/gpg-home -saq")),
             verify_cmd: Some(Command::from("gpg --homedir ./tests/gpg-home --verify -q")),