+++ mml/CHANGELOG.md
@@ -12,6 +12,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `tokio` and `async-std` cargo features (they are both mutually exclusive).
 - Added `rustls` and `native-tls` cargo features (they are both mutually exclusive)
 - Added `secret-keyring` and `secret-command` cargo features (they are forwarded to the `secret-lib` crate)
+- Added `Pgp::None` variant with associated `Error::PgpMissingConfigurationError` error.
+
+### Changed
+
+- Renamed `Pgp::Cmds` to `Pgp::Commands`.
+- Renamed `Gpg` to `PgpGpg`.
+- Renamed `CmdsPgp` to `PgpCommands`.
+- Renamed `NativePgp` to `PgpNative`.
 
 ## [1.0.14] - 2024-08-16